dodata skripta za bazu

+++ src/main/java/com/project/model/Pacijent.java
@@ -26,13 +26,13 @@ public class Pacijent extends Korisnik{
 		
 	}
 	
-	public Pacijent(Integer id, String email, String lozinka, String ime, String prezime, String adresa, String broj,String lbo, ZdravstveniKarton zdravstveniKarton) {
+	public Pacijent(Long id, String email, String lozinka, String ime, String prezime, String adresa, String broj,String lbo, ZdravstveniKarton zdravstveniKarton) {
 		super(id, email, lozinka, ime, prezime, adresa, broj);
 		this.lbo = lbo;
 	}
 	
 	public Pacijent(String email, String lozinka, String ime, String prezime, String adresa, String broj,String lbo) {
-		super(0, email, lozinka, ime, prezime, adresa, broj);
+		super(0L, email, lozinka, ime, prezime, adresa, broj);
 		this.lbo = lbo;
 	}
