+++ src/main/java/com/project/repository/PacijentRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.project.model.Pacijent;
 
-public interface PacijentRepository extends JpaRepository<Pacijent, Integer>{
+public interface PacijentRepository extends JpaRepository<Pacijent, Long>{
 
 }
